create multiple code cell
