+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx":"1","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx":"2","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\preview.tsx":"3","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\index.ts":"4","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\fetch-plugin.ts":"5","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\unpkg-path-plugin.ts":"6"},{"size":1169,"mtime":1689956723366,"results":"7","hashOfConfig":"8"},{"size":2340,"mtime":1689954374138,"results":"9","hashOfConfig":"8"},{"size":1091,"mtime":1689955426207,"results":"10","hashOfConfig":"8"},{"size":868,"mtime":1689956855985,"results":"11","hashOfConfig":"8"},{"size":3250,"mtime":1689612354887,"results":"12","hashOfConfig":"8"},{"size":2095,"mtime":1689266132893,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1adxqzh",{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx",[],["27","28"],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\preview.tsx",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\index.ts",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\fetch-plugin.ts",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\unpkg-path-plugin.ts",["29"],{"ruleId":"30","replacedBy":"31"},{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","severity":1,"message":"35","line":1,"column":8,"nodeType":"36","messageId":"37","endLine":1,"endColumn":13},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],"@typescript-eslint/no-unused-vars","'axios' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
+[{"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx":"1","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx":"2","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\preview.tsx":"3","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\index.ts":"4","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\fetch-plugin.ts":"5","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\unpkg-path-plugin.ts":"6","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-cell.tsx":"7"},{"size":575,"mtime":1689957668900,"results":"8","hashOfConfig":"9"},{"size":2340,"mtime":1689954374138,"results":"10","hashOfConfig":"9"},{"size":1091,"mtime":1689955426207,"results":"11","hashOfConfig":"9"},{"size":868,"mtime":1689956855985,"results":"12","hashOfConfig":"9"},{"size":3250,"mtime":1689612354887,"results":"13","hashOfConfig":"9"},{"size":2095,"mtime":1689266132893,"results":"14","hashOfConfig":"9"},{"size":785,"mtime":1689957747090,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1adxqzh",{"filePath":"18","messages":"19","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"fatalErrorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"29","messages":"30","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx",[],["31","32"],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\preview.tsx",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\index.ts",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\fetch-plugin.ts",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\bundler\\plugins\\unpkg-path-plugin.ts",["33"],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-cell.tsx",[],{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","severity":1,"message":"39","line":1,"column":8,"nodeType":"40","messageId":"41","endLine":1,"endColumn":13},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],"@typescript-eslint/no-unused-vars","'axios' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
